MsgPackReader: properly increment index in `ext` (#370)
* MsgPackReader: properly increment index in `ext`

The `parseExt` helper in `MsgPackReader` was not properly updating the
position in the data source. Also override `Ext.equals` to compare the
data arrays by contents instead of identity (which is consistent with
the other subclasses of `Msg`).

Added two regression tests. Note that these don't work without the
`equals` changes, since structurally identical `Ext` messages won't
compare equal unless they also happen to be the same object.

Fixes #369

* update hashcode too

* Compare serialized and re-serialized values in test

This rolls back changing `equals` and `hashCode`. Those changes may be
worth it, but not in this PR. The test compares serialized and
re-serialized values