Make JSON dictionary output of Map[K, V] usable for other types of K (#381)
Currently, only `Map[String, V]` is hardcoded to serialize as a JSON dictionary `{"...": ..., "...": ...}`, with other non-`String` `T`s serializing as a JSON list-of-lists `[["...", ...], ["...", ...]]`. The list-of-lists representation, while usable, has downsides:

1. It is much less idiomatic JSON than dictionary representation, and would be surprising for people familiar with JSON
2. It is much harder to read or skim, with lots of open and closing square brackets that all look the same, even when pretty-printed or syntax highlighted using common tools

```json
{"foo": "bar", "baz": "qux"}
[["foo", "bar"], ["baz", "qux"]]

{
  "foo": "bar",
  "baz": "qux"
}

[
  [
    "foo",
    "bar"
  ],
  [
    "baz",
    "qux"
  ]
]
```

Ideally, we should use the dictionary representation whenever possible. This PR makes it possible to use it for non-`String` types: primitives like `Int`, `Long`, `Boolean`, string-y types like `UUID` or `Duration`, and even user-defined types whose `Reader`/`Writer` inherits from `MapKey` or are wrapped in a `stringKeyRW`. For now, we leave the case where the `Reader`/`Writer` generates non-primitive JSON types like Arrays or Dictionaries as unsupported.

The change is accomplished by some tweaks to the `MapWriter0` function to remove the hardcoded special-casing of `StringWriter`, and tweaks to the `BaseElemRenderer` protocol to allow it to accept usage of non-string keys. We instead allow anything to be written when a key is required, surrounded by quotes, with the built-in quoting of strings disabled to avoid double-quoting. Again, this does not work properly for JSON Arrays or Dictionaries.

This is a forward-breaking change in the uPickle JSON serialization format: newer versions of uPickle will be able to read the old format, but older versions of uPickle will not be able to read the new format. Nevertheless, I think the problem this fixes is sufficiently important to justify the breakage. We can bump the major version number if necessary.

Also did some tightening up of the unit test suite, to make sure that the `rw` helper checks that the input value is serialized to the first of the JSON strings given. We previously did not validate this at all, only that the JSON strings all deserialized to the input value, and that the input value round tripped, but not that the string that the input value round tripped through was in the list of strings given.