Simplify enum derivation in Scala 3 (#453)
This PR allows a simple `derives` call to work on Scala 3 enums.

We add a new `macroRWAll` macro, that `derives` now forwards to. This
can be called once on each sealed trait or enum and will automatically
generate the `macroR` calls on all the sub-classes. We then add a pair
of `superTypeReader`/`superTypeWriter` forwarders to allow sub-classes
to make use of super-class readers and writers automatically without any
input from the user.

Note that this approach does not work in Scala 2; while `macroRWAll` can
be defined, `superTypeReader`/`superTypeWriter` seem to cause diverging
implicit expansions. So for now I leave the Scala 2 experience unchanged