Fix parsing of large integer doubles (#504)
Fixes https://github.com/com-lihaoyi/upickle/issues/240

The basic problem appears to be that we are using a `parseLong`
fast-path inside `ujson.Value.visitFloat64StringParts`. This is only
valid for numbers of a certain digit-length. We should be checking for
numbers larger than that length, and falling back to the `toDouble` slow
path for those