Turn back on Scala 3 tests, enable big-case-class tests (#557)
Turn on tests to verify
https://github.com/com-lihaoyi/upickle/issues/386

Scala 3 tests were accidentally disabled in
https://github.com/com-lihaoyi/upickle/pull/553, this re-enables them

The big-case-class tests do work, with the caveat you need to set
`-Xmax-inlines` to be greater than the width of your case class, and the
compilation is notably slower on 3.3.1 than in Scala 213.11. However,
the performance seems to be better in 3.4.0, and `-Xmax-inlines` is no
longer necessary.

For now, I just add `-Xmax-inlines` and accept the compilation slowness,
so as to not break anyone using Scala 3.3.1. Eventually, as people
upgrade to 3.4.0, the problem can be expected to go away